Add one-class-per-file rule to standards and refactor FakeAiClient to Fakes folder
