+++ tests/Application.Tests/Fakes/AiClientCall.cs
@@ -1,0 +1,3 @@
+namespace Application.Tests.Fakes;
+
+public record AiClientCall(string SystemPrompt, string UserPrompt);
